tweak(shared): move shared_function code out of each file
