+++ code/components/citizen-scripting-v8/src/V8ScriptRuntime.cpp
@@ -16,6 +16,7 @@
 #include <stack>
 
 #include <CoreConsole.h>
+#include <SharedFunction.h>
 
 #include <v8-version.h>
 
@@ -515,31 +516,6 @@ template<typename... TArgs>
 void ScriptTrace(const char* string, const TArgs& ... args)
 {
 	ScriptTraceV(string, fmt::make_printf_args(args...));
-}
-
-// blindly copypasted from StackOverflow (to allow std::function to store V8 UniquePersistent types with their move semantics)
-template<class F>
-struct shared_function
-{
-	std::shared_ptr<F> f;
-	shared_function() = default;
-	shared_function(F&& f_) : f(std::make_shared<F>(std::move(f_))) {}
-	shared_function(shared_function const&) = default;
-	shared_function(shared_function&&) = default;
-	shared_function& operator=(shared_function const&) = default;
-	shared_function& operator=(shared_function&&) = default;
-
-	template<class...As>
-	auto operator()(As&&...as) const
-	{
-		return (*f)(std::forward<As>(as)...);
-	}
-};
-
-template<class F>
-shared_function<std::decay_t<F>> make_shared_function(F&& f)
-{
-	return { std::forward<F>(f) };
 }
 
 enum class V8MetaFields
