+++ code/components/citizen-server-impl/src/ResourceHttpHandler.cpp
@@ -23,6 +23,7 @@
 #include <GameServerComms.h>
 
 #include <MonoThreadAttachment.h>
+#include <SharedFunction.h>
 
 #include <TcpListenManager.h>
 
@@ -33,32 +34,6 @@ static auto GetHttpHandler(fx::Resource* resource)
 	{
 
 	};
-}
-
-// blindly copypasted from StackOverflow (to allow std::function to store the funcref types with their move semantics)
-// TODO: we use this twice now, time for a shared header?
-template<class F>
-struct shared_function
-{
-	std::shared_ptr<F> f;
-	shared_function() = default;
-	shared_function(F&& f_) : f(std::make_shared<F>(std::move(f_))) {}
-	shared_function(shared_function const&) = default;
-	shared_function(shared_function&&) = default;
-	shared_function& operator=(shared_function const&) = default;
-	shared_function& operator=(shared_function&&) = default;
-
-	template<class...As>
-	auto operator()(As&&...as) const
-	{
-		return (*f)(std::forward<As>(as)...);
-	}
-};
-
-template<class F>
-shared_function<std::decay_t<F>> make_shared_function(F&& f)
-{
-	return { std::forward<F>(f) };
 }
 
 class ResourceHttpComponent : public fwRefCountable, public fx::IAttached<fx::Resource>
